Move `Experience.ts` into `ExperienceService.ts`
